Create README.md
The main objective is to enable users to get their damaged electronic device fixed at comfort of their home without necessarily having to move around from one repair shop to another. And also to enable the user to request for an expert to be send to their location to help them install electronic devices they don’t know how to fix. 
The System aims to automate the activities carried out in the current manual system of repairing damaged devices.